Link GUI to Duke's logic
