+++ src/main/java/duke/Launcher.java
@@ -7,6 +7,6 @@ import javafx.application.Application;
  */
 public class Launcher {
     public static void main(String[] args) {
-        Application.launch(Ui.class, args);
+        Application.launch(Duke.class, args);
     }
 }
